requirements.txt
asgiref==3.8.1
boto3==1.41.4
botocore==1.41.4
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
cryptography==44.0.0
dj-database-url==3.0.1
Django==5.2.8
django-allauth==65.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-storages==1.14.6
djangorestframework==3.16.1
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
idna==3.11
jmespath==1.0.1
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0